Mask Secrets automatically and Fixed error in masking logic
Fields can now be automatically masked, if the field name contains certain phrases.&#10;&#10;Also the AuditEntryGenerator did not check the annotations of any changed subobject, that gets logged. Therefore leaking secrets or not ignoring fields properly. This is now also fixed.&#10;&#10;This PR depends on the changes made in this PR: https://ecosystem.cloudogu.com/scm/repo/scm-manager/scm-manager/pull-request/184/diff/&#10;&#10;It will not build until the PR 184 of the core gets merged

Pushed-by: k8s-git-ops<admin@cloudogu.com>
Co-authored-by: Thomas Zerr<thomas.zerr@cloudogu.com>
Pushed-by: Thomas Zerr<thomas.zerr@cloudogu.com>


Reviewed-by: Konstantin Schaper <konstantin.schaper@cloudogu.com>